+++ webpack.config.js
@@ -21,4 +21,9 @@ export default {
       },
     ],
   },
+  externals: {
+    'react/addons': true,
+    'react/lib/ReactContext': true,
+    'react/lib/ExecutionEnvironment': true,
+  },
 };