+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Open Source Documents (ChangeLog, Readme, License)
 - Basic build (React, Babel, Webpack)
 - Added QA tools (ESLint)
+- Added Testing (Karma, Mocha, Chai, Enzyme)